Fix linking of ALL THE THINGS
I mistakenly indented the LDFLAGS lines in Makefile.am, which causes
automake to copy those lines literally to the makefile instead of
processing them properly, which means they were haha never actually
executed, which means we were never linking properly with the frameworks.

The only reason this worked even though we did not link against any
of the frameworks that have the symbols that we need is because
openpa was 1) linked properly and 2) static and 3) we never directly
use the missing symbols in our code, so at link time the linker did
not pick up that it was missing symbol definitions (that were basically
linked-deferred in openpa).

When running in Unity, Unity already has all the missing symbols loaded
in the process space, which meant we never had issues finding those symbols.
When running the native test-console, however, there's nothing loaded in
the process so test-console would fail at runtime because it could not
find kCF* symbols (CoreFoundation), a BIG HINT that linking was missing
key frameworks.